pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nyckel"]

[project]
name = "nyckel"
version = "0.4.2"
authors = [{ name = "Oscar Beijbom", email = "oscar@nyckel.com" }]
description = "Python package for the Nyckel API"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = ["tqdm>=4.0.0", "requests>=2.10.0", "pillow>=8.0.0"]

[project.urls]
"Homepage" = "https://github.com/NyckelAI/python-sdk"

[tool.ruff]
line-length = 120
select = ["E", "F", "I"]

[tool.black]
line-length = 120

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
warn_return_any = false
